[HEALTH] Fixed 2+ Death or Exit calls in one frame causing error in EnemyCharacterMaster.OnDeathOrExit due to m_EnemyWave being cleared in first call
This is a situation hard to reproduce as you need to kill an enemy by hitting it on the same frame with 2 projectiles or a melee attack and a projectile, with the first hit already killing the target. Alternatively killing a target on the same frame it exits the Living Zone.

HealthSystem now checks if entity is alive via IPooledObject.IsInUse, which includes health > 0 (not dead) but also that entity didn't exit Living Zone and was Released due to this.

In addition, we set Living Zone SEO after Default Time which includes the Projectile hit and Melee Attack animation event hit, so this can only be checked after hits. It was already checking for IsAlive and was safe anyway.

We still refactored this check to also use IPooledObject.IsInUse, as the m_IsAlive flag happens to be in sync with it anyway.