[COOK] Added CookedEnemy component to give unique behaviour to Cooked Enemy
Merged all levels of Cooked Enemy into one prefab able to switch sprite via animation on init and refresh.
CookedEnemy implements new interface IPickUpEffect, allowing it to customize pick-up effect. Removed PickUpParameters since each IPickUpEffect will be able to apply a custom effect with a custom parameters asset type.
Currently it acts the same, only changing on init, but later we'll plug extra fire damage. It is already on layer PickUpTangible that collides with Projectile and has a damage API that should refresh the cook level along with the sprite and effect on consumption.