[SETUP] Intention scripts are now all ClearableBehaviours and handle their own clear. Removed manual clear from controller scripts.
This was done since we added BT, a kind of super-controller, that is, according to our logic, responsible for clearing intention, but in practice it doesn't know about all the existing intention scripts that it may indirectly modify via actions; and we don't want to iterate on all actions to clear anything they may have changed, because it would be redundant if we have at least two actions affecting the same intention, which we often do.

This left a bug where intentions were sticky after Restart Level with an enemy that only used a BT, without standard controller. Now, intentions will be cleared in all cases on Restart.