[SCROLLING] Fixed Echinemon spawned soil moving at 2x expected speed
Miground Layer was moving at anti-scrolling speed with children not rendered on Background Camera causing 2x expected speed.

Now, Midground Layer is static, on EnvironmentProp layer (actually it's children should be, to be rendered on main camera not background camera). Later, it may even be merged with the gameplay environment.