[ACTION] Action: Move Along Path now takes a localOffsetFromAnchor parameter to move aside the curve, with an offset rotated by the current tangent angle
This is to prepare squad/swarm enemy waves where enemies will move along the inner/outer curve of some shared path (or at least separate copies of paths that happen to be the same thanks to enemy prefabs) but with different local offsets.

Also fixed edge case in RunUpdate when degenerated key points would slow down speed so much that we would return early but never finish the action, as the parameter would never be updated.

We now always work with anchor, then compute final position from it.

! Setting big offsets can cause very fast motion, there is no clamping on speed yet. Consider adding one to avoid crazy motion and smoothing things better !