[REFACTOR] Replaced EventEffect_SpawnProjectilesOnDirectionalDamage component on Rock prefab with new SpawnProjectilesOnDeathWithDirectionalDamage component
The latter is basically a copy of the first one (including its base class code), except it implements IDeathHandler. This way, it auto-registers to HealthSystem on the same game object and works out of the box, without needing an Event Trigger: Entity Death.

We kept the old SpawnProjectiles event effects just in case, but we can probably remove the OnDirectionalDamage version and even the non-directional version later if they are not used.